chore: rename kani::visitor to reachability as original code does
